refactor(test): remove abstract test class, replicate tests per DB engine
- Remove AbstractDatabaseSessionStorageTest (caused PHPUnit warning)
- Each engine test (SQLite, MySQL, MSSQL) is now self-contained
- No abstract class, no trait, no inheritance
- Eliminates 'abstract class' PHPUnit warning in CI

796 tests, 0 errors, 0 failures, 0 abstract class warnings.
